Split off consumer, producer, literal from hot.ts
