+++ src/common/base/consumer.ts
@@ -1,0 +1,3 @@
+/** A function that takes a value of type `T` as input. */
+export type Consumer<T> = (input: T) => void;
+
