+++ src/common/base/producer.ts
@@ -1,0 +1,3 @@
+/** A function type that returns a value of type `T` */
+export type Producer<T> = () => T;
+
