+++ src/runtime/particle.ts
@@ -11,7 +11,8 @@
 import {BigCollection} from './handle.js';
 import {Collection} from './handle.js';
 import {Handle} from './handle.js';
-import {Runnable, Consumer} from '../common/hot.js';
+import {Runnable} from '../common/hot.js';
+import {Consumer} from '../common/base/consumer.js';
 import {InnerArcHandle} from './particle-execution-context.js';
 import {HandleConnectionSpec, ParticleSpec} from './particle-spec.js';
 import {Relevance} from './relevance.js';
