+++ src/runtime/storage/storage-provider-base.ts
@@ -17,7 +17,7 @@ import {ModelValue, SerializedModelEntry} from './crdt-collection-model.js';
 import {KeyBase} from './key-base.js';
 import {Store, BigCollectionStore, CollectionStore, SingletonStore} from '../store.js';
 import {PropagatedException} from '../arc-exceptions.js';
-import {Consumer} from '../../common/hot.js';
+import {Consumer} from '../../common/base/consumer.js';
 import {Dictionary} from '../../common/collect/dictionary.js';
 import {ClaimIsTag} from '../particle-claim.js';
