+++ src/runtime/type.ts
@@ -15,7 +15,7 @@ import {Schema} from './schema.js';
 import {SlotInfo} from './slot-info.js';
 import {ArcInfo} from './synthetic-types.js';
 import {TypeVariableInfo} from './type-variable-info.js';
-import {Literal} from '../common/hot.js';
+import {Literal} from '../common/base/literal.js';
 import {Predicate} from '../common/base/predicate.js';
 
 export interface TypeLiteral extends Literal {
